Moved some of the rtk stuff out of radar.html. Seperated the code into the correct files
